split module_sources.py into app_source_connections.py (loading remote app data) and app_loading.py (loading app data into the database)
