+++ guidedmodules/management/commands/apps_catalog.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand, CommandError
 from django.conf import settings
 
 from guidedmodules.models import AppSource
-from guidedmodules.module_sources import MultiplexedAppSourceConnection
+from guidedmodules.app_source_connections import MultiplexedAppSourceConnection
 
 class Command(BaseCommand):
     help = 'Prints a list of all apps in the configured app stores.'
